+++ examples/advanced/src/gen/zod/petSchema.ts
@@ -4,10 +4,10 @@ import { categorySchema } from './categorySchema'
 import { tagSchema } from './tagSchema'
 
 export const petSchema = z.object({
-  id: z.number().optional(),
+  id: z.number().readonly().optional(),
   name: z.string(),
-  category: z.lazy(() => categorySchema).optional(),
+  category: z.lazy(() => categorySchema).schema.optional(),
   photoUrls: z.array(z.string()),
-  tags: z.array(z.lazy(() => tagSchema)).optional(),
+  tags: z.array(z.lazy(() => tagSchema).schema).optional(),
   status: z.enum([`available`, `pending`, `sold`]).describe(`pet status in the store`).optional(),
 })