+++ examples/faker/src/gen/mocks/createCustomer.ts
@@ -3,6 +3,6 @@ import { faker } from '@faker-js/faker'
 import { createAddress } from './createAddress'
 import { Customer } from '../models/Customer'
 
-export function createCustomer(): Customer {
+export function createCustomer(): NonNullable<Customer> {
   return { id: faker.number.float({}), username: faker.string.alpha(), address: faker.helpers.arrayElements([createAddress()]) as any }
 }