+++ examples/msw/src/gen/mocks/createCategory.ts
@@ -2,6 +2,6 @@ import { faker } from '@faker-js/faker'
 
 import { Category } from '../models/Category'
 
-export function createCategory(): Category {
+export function createCategory(): NonNullable<Category> {
   return { id: faker.number.float({}), name: faker.string.alpha() }
 }