+++ examples/msw/src/gen/mocks/petMocks/createAddPet.ts
@@ -6,7 +6,7 @@ import { AddPet405 } from '../../models/AddPet'
 import { AddPetMutationRequest } from '../../models/AddPet'
 import { AddPetMutationResponse } from '../../models/AddPet'
 
-export function createAddPet405(): AddPet405 {
+export function createAddPet405(): NonNullable<AddPet405> {
   return { code: faker.number.float({}), message: faker.string.alpha() }
 }
 
@@ -14,7 +14,7 @@ export function createAddPet405(): AddPet405 {
  * @description Create a new pet in the store
  */
 
-export function createAddPetMutationRequest(): AddPetMutationRequest {
+export function createAddPetMutationRequest(): NonNullable<AddPetMutationRequest> {
   return createAddPetRequest()
 }
 
@@ -22,6 +22,6 @@ export function createAddPetMutationRequest(): AddPetMutationRequest {
  * @description Successful operation
  */
 
-export function createAddPetMutationResponse(): AddPetMutationResponse {
+export function createAddPetMutationResponse(): NonNullable<AddPetMutationResponse> {
   return createPet()
 }