+++ examples/zod/src/gen/zod/createUsersWithListInputSchema.ts
@@ -6,9 +6,9 @@ import { userSchema } from './userSchema'
  * @description successful operation
  */
 export const createUsersWithListInputErrorSchema = z.any()
-export const createUsersWithListInputMutationRequestSchema = z.array(z.lazy(() => userSchema))
+export const createUsersWithListInputMutationRequestSchema = z.array(z.lazy(() => userSchema).schema)
 
 /**
  * @description Successful operation
  */
-export const createUsersWithListInputMutationResponseSchema = z.lazy(() => userSchema)
+export const createUsersWithListInputMutationResponseSchema = z.lazy(() => userSchema).schema