+++ examples/zod/src/gen/zod/userArraySchema.ts
@@ -2,4 +2,4 @@ import { z } from 'zod'
 
 import { userSchema } from './userSchema'
 
-export const userArraySchema = z.array(z.lazy(() => userSchema))
+export const userArraySchema = z.array(z.lazy(() => userSchema).schema)