+++ packages/swagger-faker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.11.5](https://github.com/kubb-project/kubb/compare/@kubb/swagger-faker-v1.11.4...@kubb/swagger-faker-v1.11.5) (2023-10-16)
+
+
+### 📦 Miscellaneous Chores
+
+* **@kubb/swagger-faker:** Synchronize undefined versions
+
 ## [1.11.4](https://github.com/kubb-project/kubb/compare/@kubb/swagger-faker-v1.11.3...@kubb/swagger-faker-v1.11.4) (2023-10-16)